% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DAG.R
\name{DAG}
\alias{DAG}
\title{A DAG Function}
\usage{
DAG(X, K, tau, B = NULL, A0 = NULL, opts.tol = 1e-05,
  equal.sigma = FALSE)
}
\arguments{
\item{X}{The data matrix}

\item{K}{tuning parameter}

\item{tau}{tuning parameter}

\item{B}{A p by p matrix indicating nonzero elements as initial values}

\item{A0}{A p by p matrix as initial values for A}

\item{opts.tol}{Tolerance for convergence}

\item{equal.sigma}{Whether equal error variance is assumed.}
}
\value{
Estimated adjacency matrix and Estimated error variances for each node
}
\description{
This function allows you to learn the DAG structure from Gaussian data
}
\examples{
m=50
amat <- matrix(0, m, m)
amat[2:m,1]<- 1 #(rbinom((m-1),1,0.5)-0.5)*runif((m-1),0.5,2)
sigma <- seq(1,0.5,length.out=m)
X <- rmvDAG(n,amat,sigma)
out<-DAG(X,10,0.01)
}
\keyword{DAG}

