% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvDAG.R
\name{rmvDAG}
\alias{rmvDAG}
\title{A rmvDAG Function}
\usage{
rmvDAG(n, amat, sigma = NULL)
}
\arguments{
\item{n}{The sample size}

\item{amat}{The adjacency matrix of a DAG}

\item{sigma}{The error variance of each node}
}
\value{
Gaussian data with the given sample size
}
\description{
This function allows you to generate Gaussian data from a DAG
}
\examples{
amat=matrix(c(0,1,0,0),2,2)
rmvDAG(50,amat)
}
\keyword{rmvDAG}
