% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained.R
\name{uncon_dag}
\alias{uncon_dag}
\title{A DAG Function}
\usage{
uncon_dag(X, lambda, tau, B = NULL, A0 = NULL, opts.tol = 1e-04)
}
\arguments{
\item{X}{The data matrix}

\item{lambda}{tuning parameter for the unconstrained version}

\item{tau}{tuning parameter}

\item{B}{A p by p matrix indicating nonzero elements as initial values}

\item{A0}{A p by p matrix as initial values for A}

\item{opts.tol}{Tolerance for convergence}
}
\value{
Estimated adjacency matrix and Estimated error variances for each node
}
\description{
This function allows you to learn the DAG structure from Gaussian data
}
\examples{
m=50
amat <- matrix(0, m, m)
amat[2:m,1]<- 1 #(rbinom((m-1),1,0.5)-0.5)*runif((m-1),0.5,2)
sigma <- seq(1,0.5,length.out=m)
X <- rmvDAG(n,amat,sigma)
out<-uncon_dag(X,5,0.01)
}
\keyword{DAGun}
